+++ src/ConfigurableFirmata.h
@@ -181,7 +181,7 @@ class FirmataClass
     void setPinMode(byte pin, byte config);
     /* access pin state */
     int getPinState(byte pin);
-    void setPinState(byte pin, int state);
+    void setPinState(byte pin, byte state);
 
     /* utility methods */
     void sendValueAsTwo7bitBytes(int value);
@@ -203,7 +203,7 @@ class FirmataClass
     int sysexBytesRead;
     /* pins configuration */
     byte pinConfig[TOTAL_PINS];         // configuration of every pin
-    int pinState[TOTAL_PINS];           // any value that has been written
+    byte pinState[TOTAL_PINS];           // any value that has been written
 
     boolean resetting;
